Fix time pretty-printing

+++ foobnix/util/time_utils.py
@@ -17,13 +17,13 @@ def size2text(size):
 def convert_seconds_to_text(time_sec):
         time_sec = int(time_sec)
 
-        hours = time_sec / (60 * 60)
+        hours = int(time_sec / (60 * 60))
         time_sec = time_sec - (hours * 60 * 60)
 
-        mins = time_sec / 60
+        mins = int(time_sec / 60)
         time_sec = time_sec - (mins * 60)
 
-        secs = time_sec
+        secs = int(time_sec)
         if hours > 0:
             return '%(hours)d:%(mins)02d:%(secs)02d' % {'hours' : hours, 'mins': mins, 'secs': secs }
         else:
